Tafelbild 1.0.3: Substring-Fehler in NameList.parse behoben
`String.split` liefert Substrings; die App-eigene Erweiterung `trimmed`
gibt es nur auf `String`. Xcode meldete deshalb „Reference to member
'trimmed' cannot be resolved without a contextual type".

- NameList.parse wandelt jeden Teil erst in String, dann wird getrimmt.
- Vorbeugend alle Rechnungen entzerrt, die Double (Modell, Tafelpunkte)
  und CGFloat (Ansicht) mischen: Layout hat jetzt eigene Double-Maße
  (canvasWidth/-Height, minWidth/-Height), die Zieh- und Größengesten,
  die Platzsuche für neue Elemente, die Stapelansicht sowie einzelne
  Balken- und Skalenbreiten rechnen ausdrücklich um.
- Version 1.0.3.

+++ TafelbildiOS/Tafelbild/Views/Sheets/WidgetSettingsSheet.swift
@@ -394,7 +394,7 @@ private struct NoiseSettings: View {
                         Capsule().fill(Color.primary.opacity(0.1))
                         Capsule()
                             .fill(meter.level * content.gain > content.threshold ? Theme.danger : Theme.mint)
-                            .frame(width: geo.size.width * min(1, meter.level * content.gain))
+                            .frame(width: geo.size.width * CGFloat(min(1, meter.level * content.gain)))
                     }
                 }
                 .frame(height: 8)
